^ y_cooldown (from 255 to 65535)

+++ src/common/player-common.h
@@ -938,7 +938,7 @@ struct player
     char account_name[NORMAL_WID];  // account name (needed for housing)
     uint32_t account_score;         // account score (for ladder)
     uint8_t supporter;              // supporter's account lvl
-    uint8_t y_cooldown;             // cooldown for racial abilities
+    uint16_t y_cooldown;            // cooldown for racial abilities
     int16_t iron_timer;             // time until next auto > for ironman
 
     // Weather
